add more decorated info when logging error messages
