+++ src/utils/error-logger.js
@@ -1,9 +1,10 @@
-const { createLogger, format, transports } = require('winston');
+const { createLogger, transports } = require('winston');
+const customFormat = require('./logger-custom-format');
 
 const errLogger = createLogger({
   level: 'error',
   transports: new transports.Console(),
-  format: format.simple(),
+  format: customFormat(),
 });
 
 module.exports = errLogger;
